refactor - move docstring code out of server package to avoid circ. dependencies
